pyproject.toml
[project]
name = "youtube-playlist-mcp"
version = "0.1.0"
description = "MCP server for managing YouTube playlists via the YouTube Data API v3"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[project.scripts]
youtube-playlist-mcp = "youtube_playlist_mcp.server:main"
authenticate = "youtube_playlist_mcp.authenticate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_playlist_mcp"]
